requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
requests
streamlit
matplotlib
numpy
gTTS
